separate rpc server into a provider mod
