+++ src/lib.rs
@@ -2,6 +2,7 @@ pub mod bitcoin_utils;
 pub mod ipc_lib;
 pub mod ipc_state;
 pub mod l1_manager;
+pub mod provider;
 pub mod subnet_simulator;
 pub mod utils;
